requirements.txt
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
gcloud==0.17.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jws==0.1.3
oauth2client==3.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
